Move extra constants out, move functions out
