+++ src/backend/config.ts
@@ -5,7 +5,7 @@ import {
   GlobalConfigVersion,
   WineInstallation
 } from 'common/types'
-import { currentGlobalConfigVersion, getSteamCompatFolder } from './constants'
+import { currentGlobalConfigVersion } from 'backend/constants/others'
 
 import { logError, logInfo, LogPrefix } from './logger/logger'
 import {
@@ -34,6 +34,36 @@ import {
   heroicInstallPath,
   userHome
 } from './constants/paths'
+import { join } from 'path'
+import { spawnSync } from 'child_process'
+
+function getSteamCompatFolder() {
+  // Paths are from https://savelocation.net/steam-game-folder
+  if (isWindows) {
+    const defaultWinPath = join(process.env['PROGRAMFILES(X86)'] ?? '', 'Steam')
+    return defaultWinPath
+  } else if (isMac) {
+    return join(userHome, 'Library/Application Support/Steam')
+  } else {
+    const flatpakSteamPath = join(
+      userHome,
+      '.var/app/com.valvesoftware.Steam/.steam/steam'
+    )
+
+    if (existsSync(flatpakSteamPath)) {
+      // check if steam is really installed via flatpak
+      const { status } = spawnSync('flatpak', [
+        'info',
+        'com.valvesoftware.Steam'
+      ])
+
+      if (status === 0) {
+        return flatpakSteamPath
+      }
+    }
+    return join(userHome, '.steam/steam')
+  }
+}
 
 /**
  * This class does config handling.
