+++ src/backend/tools/index.ts
@@ -20,7 +20,7 @@ import {
   extractFiles,
   getWineFromProton
 } from '../utils'
-import { execOptions } from '../constants'
+import { execOptions } from 'backend/constants/others'
 import { logError, logInfo, LogPrefix, logWarning } from '../logger/logger'
 import i18next from 'i18next'
 import { dirname, join } from 'path'
