+++ src/backend/utils.ts
@@ -19,13 +19,13 @@ import {
   SpawnOptions,
   spawnSync
 } from 'child_process'
-import { existsSync, rmSync } from 'graceful-fs'
+import { existsSync, readFileSync, rmSync } from 'graceful-fs'
 import { promisify } from 'util'
 import i18next, { t } from 'i18next'
 
-import { getSteamLibraries } from './constants'
 import {
   appendGamePlayLog,
+  logDebug,
   logError,
   logInfo,
   LogPrefix,
@@ -85,6 +85,7 @@ import {
   publicDir,
   windowIcon
 } from './constants/paths'
+import { parse } from '@node-steam/vdf'
 
 const execAsync = promisify(exec)
 
@@ -518,6 +519,29 @@ function getFormattedOsName(): string {
     default:
       return 'Unknown OS'
   }
+}
+
+export async function getSteamLibraries(): Promise<string[]> {
+  const { defaultSteamPath } = GlobalConfig.get().getSettings()
+  const path = defaultSteamPath.replaceAll("'", '')
+  const vdfFile = join(path, 'steamapps', 'libraryfolders.vdf')
+  const libraries = ['/usr/share/steam']
+
+  if (existsSync(vdfFile)) {
+    const json = parse(readFileSync(vdfFile, 'utf-8'))
+    if (!json.libraryfolders) {
+      return libraries
+    }
+    const folders: { path: string }[] = Object.values(json.libraryfolders)
+    return [...libraries, ...folders.map((folder) => folder.path)].filter(
+      (path) => existsSync(path)
+    )
+  }
+  logDebug(
+    'Unable to load Steam Libraries, libraryfolders.vdf not found',
+    LogPrefix.Backend
+  )
+  return libraries
 }
 
 async function getSteamRuntime(
